% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.calcf.r
\name{am.calcf}
\alias{am.calcf}
\title{Calculate new features}
\usage{
am.calcf(ds, ds.list, verbose = FALSE)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{data.table or character vector containing new columns temporary names obtained with am.scor of am.scorfl functions. If in the vector format names should be formatted in this way: oper***feature1_name***feature2_name. For example: mult***f10***f25. Data.table should contain character columns oper, f1, f2.}

\item{verbose}{logical}
}
\value{
Data.table. rn - new feature name, cr - correlation, f1 - first original feature used, f2 - second original feature used
}
\description{
Calculate new features based on the vector of specially formatted column pairs
}
\seealso{
am.scor, am.scorfl
}

