% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.checkbiny.r
\name{am.checbkbiny}
\alias{am.checbkbiny}
\title{Get correlation between predicted Y variable and residuals splitted per binary variables}
\usage{
am.checbkbiny(ds, ds.list, ds.y, ds.resid, corr.type = "spearman")
}
\arguments{
\item{ds}{data.table: data set. Must include binary variables we want to test, residuals and the predicted Y variable}

\item{ds.list}{character vector: names of independent binary variables (X)}

\item{ds.y}{character: name of the predicted dependent variable (Y)}

\item{ds.resid}{character: name of the residuals variable}

\item{corr.type}{character: correlation types; "spearman" (default) and "pearson" available; see Hmisc rcorr for details}
}
\value{
Data.table. rn - feature name, cr - correlation value, prc - per cent of of the whole dataset, vl - value of splitting variable
}
\description{
Imagine we want to test if residuals when c1 variable == 1 have stronger correlation to predicted variable than when c1 == 0.
If it is the case, use this functions.
}
\seealso{
...
}

