% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.checkx.r
\name{am.checkx}
\alias{am.checkx}
\title{Autocorrelations between independent variables}
\usage{
am.checkx(ds, ds.list, corr.type = "spearman", cl.number = 1, sns = 0.5,
  out.file = "")
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector: names of independent variables (X)}

\item{corr.type}{character: correlation types; "spearman" (default) and "pearson" available; see Hmisc rcorr for details}

\item{cl.number}{number: cluster number for parallel computers; be very careful with this parameter! do not set it too big}

\item{sns}{number: Sensitivity. How strong should be probable intersectons between two features to test autocorrelations}

\item{out.file}{character: absolute or relative path to the files with output; due to parallel computations you do NOT see most of the info in the console; id default no outputs are used}
}
\value{
Data.table. rn - feature name, cr - correlation value
}
\description{
Define pairwise correlations between independent variables. Use sns argument to define what correlation you want to test. Be careful! Parallel computation in use.
}
\seealso{
You can use this data to choose and create new features you prefer with am.calcf function
}

