% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.checky.r
\name{am.checky}
\alias{am.checky}
\title{Correlations between dependent variable (Y) and every independent variable}
\usage{
am.checky(ds, ds.list, ds.y, corr.type = "spearman", cl.number = 1,
  out.file = "")
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.y}{character: name of the dependent variable (Y)}

\item{corr.type}{character: correlation types; "spearman" (default) and "pearson" available; see Hmisc rcorr for details}

\item{cl.number}{number: cluster number for parallel computers; be very careful with this parameter! do not set it too big}

\item{out.file}{character: absolute or relative path to the files with output; due to parallel computations you do NOT see most of the info in the console; id default no outputs are used}
}
\value{
Data.table. rn - feature name, cr - correlation value
}
\description{
Define pairwise correlations between the dependant variable (Y) and every independent variable (Xs). Be careful! Parallel computation in use.
}
\seealso{
You can use this data to choose and create new features you prefer with am.calcf function
}

