% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.conv.r
\name{amm.conv}
\alias{amm.conv}
\title{Convert columns to appropriate type}
\usage{
amm.conv(ds, ds.list)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}
}
\value{
Data.table
}
\description{
Convert columns to appropriate type. Needs more memory, does not modify the original data set
}
\seealso{
amm.convl for more memory efficient converting
}

