% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.conv.r
\name{amm.convl}
\alias{amm.convl}
\title{Convert columns to appropriate type (pre link)}
\usage{
amm.convl(ds, ds.list)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}
}
\value{
Data.table
}
\description{
Convert columns to appropriate type per link. Needs less memory, modifies the original data set
}
\seealso{
amm.conv for converting without modifying the original data set
}

