% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.factor.r
\name{amm.factor}
\alias{amm.factor}
\title{Set factors}
\usage{
amm.factor(ds, ds.list)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}
}
\value{
Data.table
}
\description{
Sets factors for selected columns of data.table
}
\seealso{
amm.convn for converting without modifying the original data set
}

