% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.factor.r
\name{amm.factorr}
\alias{amm.factorr}
\title{Reset factor levels}
\usage{
amm.factorr(ds)
}
\arguments{
\item{ds}{data.table: data set}
}
\value{
Data.table
}
\description{
Resets factor levels for selected columns of data.table. Particularly important for dropping unused levels
}
\seealso{
amm.factor for setting factors for selected columns
}

