% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.gbetween.r
\name{amm.gbetween}
\alias{amm.gbetween}
\title{Get numeric columns between min and max}
\usage{
amm.gbetween(ds, ds.list, excl = "", min.v = 0, max.v = 1, ptrn)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}

\item{excl}{numeric: values to exclude. NA and NaN are also possible}

\item{min.v}{numeric: min value}

\item{max.v}{numeric: max.value}
}
\value{
Character vector of column names
}
\description{
Get numeric columns contained all values between min and max. Please send only numeric columns name as ds.list param
}
\seealso{
amm.gbetweens
}

