% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.gmatch.r
\name{amm.gmatch}
\alias{amm.gmatch}
\title{Get columns by match}
\usage{
amm.gmatch(ds, ds.list, ptrn)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}

\item{ptrn}{character: pattern to find}
}
\value{
Character vector of column names
}
\description{
Get column by regexp match. Please send only character columns name as ds.list param
}
\seealso{
amm.gtmatcht
}

