% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.gmatch.r
\name{amm.gmatcht}
\alias{amm.gmatcht}
\title{Get columns by matching all the values}
\usage{
amm.gmatcht(ds, ds.list, excl = "", ptrn)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}

\item{excl}{character: values to exclude}

\item{ptrn}{character: pattern to find}
}
\value{
Character vector of column names
}
\description{
Get column by regexp matching all the values of the column. Please send only character columns name as ds.list param
}
\seealso{
amm.gmatch
}

