% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.gna.r
\name{amm.gna}
\alias{amm.gna}
\title{Get columns containing NA}
\usage{
amm.gna(ds, ds.list)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}
}
\value{
Character vector of column names
}
\description{
Get columns containing NA values
}
\seealso{
amm.gnat
}

