% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.gna.r
\name{amm.gnat}
\alias{amm.gnat}
\title{Get full NA columns}
\usage{
amm.gnat(ds, ds.list)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}
}
\value{
Character vector of column names
}
\description{
Get columns containing only NA values
}
\seealso{
amm.gna
}

