% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.gonevalue.r
\name{amm.gonevalue}
\alias{amm.gonevalue}
\title{Get columns containing only one value}
\usage{
amm.gonevalue(ds, ds.list)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}
}
\value{
Character vector of column names
}
\description{
Get columns containing only one value (including NAs)
}
\seealso{
amm.gonevaluena
}

