% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.gvalue.r
\name{amm.gvalue}
\alias{amm.gvalue}
\title{Get column by value}
\usage{
amm.gvalue(ds, ds.list, value)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}

\item{value:}{value in one of atomic types}
}
\value{
Character vector of column names
}
\description{
Get column by value (using strict type equality)
}
\seealso{
amm.match, amm.gbetween
}

