% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.remove.r
\name{amm.remove}
\alias{amm.remove}
\title{Replace column values}
\usage{
amm.remove(ds, ds.list)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}
}
\value{
Character vector of column names
}
\description{
Replace column values equal to arguments
}
\seealso{
amm.match, amm.gbetween
}

