% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.replace.r
\name{amm.replace}
\alias{amm.replace}
\title{Replace column values}
\usage{
amm.replace(ds, ds.list, ds.string, ds.replace)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}

\item{ds.string}{character vector: value to replace}

\item{ds.replace}{character: replacing character}
}
\value{
Character vector of column names
}
\description{
Replace column values equal to arguments
}
\seealso{
amm.match, amm.gbetween
}

