% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.replace.r
\name{amm.replacem}
\alias{amm.replacem}
\title{Replace columns by match}
\usage{
amm.replacem(ds, ds.list, ds.string, ds.replace)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}

\item{ds.string}{character: value to replace}

\item{ds.replace}{character: replacing character}
}
\value{
Character vector of column names
}
\description{
Replace columns by matching. Replaces the whole value
}
\seealso{
amm.match, amm.gbetween
}

