% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.replace.r
\name{amm.replacemp}
\alias{amm.replacemp}
\title{Replace pattern in column by match}
\usage{
amm.replacemp(ds, ds.list, ds.string, ds.replace)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}

\item{ds.string}{character: value to replace}

\item{ds.replace:}{replacing character}
}
\value{
Character vector of column names
}
\description{
Replace pattern in column by matching. Replaces only pattern, not the whole value
}
\seealso{
amm.match, amm.gbetween
}

