% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.torange.r
\name{amm.torange}
\alias{amm.torange}
\title{Categorical features to range ones}
\usage{
amm.torange(ds, target.col, ds.list, ...)
}
\arguments{
\item{ds}{data.table: data set}

\item{target.col}{character: Y variable for getting right order. If numeric, you can round it to some reasonable value}

\item{ds.list}{character vector of columns name needed to convert}

\item{...}{some other arguments: what ones??}
}
\value{
List of modified data set and modified ... argument
}
\description{
Convert categorical features to range ones.
}
\seealso{
...
}

