% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.torange.r
\name{amm.torange}
\alias{amm.torange}
\title{Categorical features to range ones}
\usage{
amm.torange(ds, target.col, ds.list, ...)
}
\arguments{
\item{ds}{data.table: data set}

\item{ds.list}{character vector of columns name needed to convert}

\item{...}{some other arguments: what ones??}

\item{ds.replace:}{replacing character}
}
\value{
List of modified data set and modified ... argument
}
\description{
Convert categorical features to range ones. Warning! May contain some not self evident prerequisits.
}
\seealso{
...
}

